+++ lib/pages/homePage.dart
@@ -34,8 +34,8 @@ class HomePage extends StatefulWidget {
 
   final AppService service;
   final NetworkParams connectedNode;
-  final Future<void> Function(BuildContext, PolkawalletPlugin,
-      {bool needReload}) checkJSCodeUpdate;
+  final Future<void> Function(BuildContext, PolkawalletPlugin)
+      checkJSCodeUpdate;
   final Future<void> Function(String,
       {NetworkParams node, PageRouteParams pageRoute}) switchNetwork;