+++ lib/utils/i18n/zh/public.dart
@@ -125,4 +125,5 @@ const Map<String, String> zhPublic = {
   'bridge.address': '收款地址',
   'bridge.address.warn': '不建议修改跨链转账的收款地址。\n该功能仅限高级用户使用。',
   'bridge.connecting.warn': '网络连接超时，\n请稍后重试',
+  'bridge.search.chain': '搜索链',
 };