+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# [3.3.6-beta] - 20220926
+1. Fix known bugs.
+2. UX enhancement.
+3. bump sdk/ui v0.5.0.
+
+# [3.3.5-beta] - 20220817
+1. Multiple fiat display #125
+2. 发送交易失败错误信息解析修复 #341
+3. Astar 跨链 #353
+4. polkadot-js update 8.x -> 9.x #361
+5. Swap预估收到显示优化 #359
+6. Bridge-点击头像直接进入select account页面（且UI默认选中当前账号） #347
+7. Optimize the LP assets page #242
+8. Price trend chart UI adjust #366
+9. homa taiga apy 展示调整
+
 # [3.3.4-beta] - 20220804
 1. Replace history api #267
 2. Optimize Vault display #180